added tests and correct alert EditText
