+++ app/src/androidTest/java/com/github/steroidteam/todolist/ListSelectionActivityTest.java
@@ -14,13 +14,21 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import static androidx.test.espresso.Espresso.onView;
+import static androidx.test.espresso.action.ViewActions.clearText;
 import static androidx.test.espresso.action.ViewActions.click;
 import static androidx.test.espresso.action.ViewActions.closeSoftKeyboard;
+import static androidx.test.espresso.action.ViewActions.longClick;
+import static androidx.test.espresso.action.ViewActions.replaceText;
 import static androidx.test.espresso.action.ViewActions.typeText;
 import static androidx.test.espresso.assertion.ViewAssertions.matches;
+import static androidx.test.espresso.intent.Intents.intended;
+import static androidx.test.espresso.intent.matcher.IntentMatchers.hasComponent;
+import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
+import static androidx.test.espresso.matcher.ViewMatchers.withClassName;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.matcher.ViewMatchers.withText;
 import static org.hamcrest.Matchers.anything;
+import static org.hamcrest.Matchers.endsWith;
 
 @RunWith(AndroidJUnit4.class)
 public class ListSelectionActivityTest {
@@ -31,16 +39,51 @@ public class ListSelectionActivityTest {
 
     @Test
     public void openListWorks() {
-        /**
         Intents.init();
+        Intent listSelectionActivity = new Intent(ApplicationProvider.getApplicationContext(), ListSelectionActivity.class);
 
-        // Hit the button to create a new task.
-        onView(withId(R.id.activity_list_selection_itemlist))
-                .perform(click());
-
-        Intents.intended(Matchers.allOf(IntentMatchers.hasComponent(ItemViewActivity.class.getName())));
+        try (ActivityScenario<ListSelectionActivity> scenario = ActivityScenario.launch(listSelectionActivity)) {
+            Espresso.onData(anything()).inAdapterView(withId(R.id.activity_list_selection_itemlist)).atPosition(0).perform(click());
+            intended(hasComponent(ItemViewActivity.class.getName()));
+        }
         Intents.release();
-         **/
+    }
+
+    @Test
+    public void cannotRenameTodoListWithoutText() {
+        Intent listSelectionActivity = new Intent(ApplicationProvider.getApplicationContext(), ListSelectionActivity.class);
+
+        try (ActivityScenario<ItemViewActivity> scenario = ActivityScenario.launch(listSelectionActivity)) {
+            Espresso.onData(anything()).inAdapterView(withId(R.id.activity_list_selection_itemlist)).atPosition(0).perform(longClick());
+
+            onView(withText("Please enter a new name")).check(matches(isDisplayed()));
+
+            onView(withClassName(endsWith("EditText"))).perform(clearText());
+
+            onView(withText("Confirm")).perform(click());
+
+            onView(withText("The name shouldn't be empty !")).inRoot(new ToastMatcher()).check(matches(isDisplayed()));
+        }
+    }
+
+    @Test
+    public void renameTodoListWorks() {
+        final String TODO_LIST_NAME = "Sweng Test";
+
+        Intent listSelectionActivity = new Intent(ApplicationProvider.getApplicationContext(), ListSelectionActivity.class);
+
+        try (ActivityScenario<ItemViewActivity> scenario = ActivityScenario.launch(listSelectionActivity)) {
+            Espresso.onData(anything()).inAdapterView(withId(R.id.activity_list_selection_itemlist)).atPosition(0).perform(longClick());
+
+            onView(withText("Please enter a new name")).check(matches(isDisplayed()));
+
+            onView(withClassName(endsWith("EditText"))).perform(replaceText(TODO_LIST_NAME), closeSoftKeyboard());
+
+            onView(withText("Confirm")).perform(click());
+
+            onView(withText("Successfully changed the name !")).inRoot(new ToastMatcher()).check(matches(isDisplayed()));
+            
+        }
     }
 
 }
