+++ app/src/androidTest/java/com/github/steroidteam/todolist/ToastMatcher.java
@@ -1,0 +1,31 @@
+package com.github.steroidteam.todolist;
+
+import android.os.IBinder;
+import android.view.WindowManager;
+
+import androidx.test.espresso.Root;
+
+import org.hamcrest.Description;
+import org.hamcrest.TypeSafeMatcher;
+
+public class ToastMatcher extends TypeSafeMatcher<Root> {
+
+    @Override
+    public void describeTo(Description description) {
+        description.appendText("is toast");
+    }
+
+    @Override
+    public boolean matchesSafely(Root root) {
+        int type = root.getWindowLayoutParams().get().type;
+        if (type == WindowManager.LayoutParams.TYPE_TOAST) {
+            IBinder windowToken = root.getDecorView().getWindowToken();
+            IBinder appToken = root.getDecorView().getApplicationWindowToken();
+            if (windowToken == appToken) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+}
